requirements.txt
anthropic==0.37.1
beautifulsoup4==4.12.3
datasets==2.21.0
google_search_results==2.4.2
huggingface_hub==0.23.4
langchain==0.3.4
langchain_core==0.3.13
mammoth==1.8.0
markdownify==0.13.1
numexpr==2.10.1
numpy==2.1.2
openai==1.52.2
pandas==2.2.3
pathvalidate==3.2.1
pdfminer==20191125
pdfminer.six==20240706
Pillow==11.0.0
puremagic==1.28
pypdf==5.1.0
python-dotenv==1.0.1
python_pptx==1.0.2
Requests==2.32.3
serpapi==0.1.5
tqdm==4.65.0
tqdm==4.66.4
torch==2.2.2
torchvision==0.17.2
transformers==4.46.0
youtube_transcript_api==0.6.2
